requirements.txt
opencv-python
mediapipe
pygame
numpy
Pillow
